Prepare to switch to real config method
