+++ src/NServiceBus.Gateway.AcceptanceTests/When_a_persistence_does_not_support_gateway_dedup.cs
@@ -23,7 +23,7 @@
         {
             public Endpoint()
             {
-                EndpointSetup<DefaultServer>(c =>
+                EndpointSetup<DefaultServerWithNoStorage>(c =>
                 {
                     c.EnableGateway(new GatewayConfig
                     {
@@ -35,7 +35,7 @@
                                 ChannelType = "http"
                             }
                         }
-                    }, configureInMemoryPersistence: false);
+                    });
                 });
             }
         }
